Get working with Docker
In order to get this working with Docker I had to downgrade the version
of phoenix_ecto and phoenix_html that were installed. I could not
connect to the database, and instead of failing with a sane error, it
just went into an endless loop of trying to connect and getting denied.
This was dumb.